fix warning
git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@5060 b38c6c8f-bee8-465d-8448-9abe6915d496

+++ tracking/modules/genfitter/src/GenFitterModule.cc
@@ -505,7 +505,7 @@ void GenFitterModule::event()
               //MH: this is new stuff...
               tracks[trackCounter]->setTrackFitResultIndex(chargedStable, trackFitResultCounter);
               TrackFitResult* newTrackFitResult = trackFitResults.appendNew();
-              newTrackFitResult->setCharge(gfTrack.getCharge());
+              newTrackFitResult->setCharge((int)gfTrack.getCharge());
               newTrackFitResult->setMomentum(resultMomentum);
               newTrackFitResult->setPosition(resultPosition);
               newTrackFitResult->setCovariance6(newResultCovariance);
